Move CommentsHeader to it's own file
