+++ osu.Game.Tests/Visual/Online/TestSceneCommentsContainer.cs
@@ -17,6 +17,8 @@ namespace osu.Game.Tests.Visual.Online
         public override IReadOnlyList<Type> RequiredTypes => new[]
         {
             typeof(CommentsContainer),
+            typeof(CommentsHeader),
+            typeof(DrawableComment),
         };
 
         protected override bool UseOnlineAPI => true;
