+++ osu.Game/Overlays/Comments/CommentsContainer.cs
@@ -9,8 +9,6 @@ using osu.Framework.Graphics;
 using osu.Framework.Bindables;
 using osu.Framework.Graphics.Shapes;
 using osu.Game.Graphics;
-using osu.Framework.Graphics.Sprites;
-using osuTK;
 using osu.Game.Online.API.Requests.Responses;
 
 namespace osu.Game.Overlays.Comments
@@ -114,60 +112,6 @@ namespace osu.Game.Overlays.Comments
         private void load()
         {
             background.Colour = colours.Gray3;
-        }
-
-        private class CommentsHeader : CompositeDrawable
-        {
-            private const int height = 40;
-            private const int spacing = 10;
-            private const int padding = 50;
-
-            public readonly Bindable<SortCommentsBy> Sort = new Bindable<SortCommentsBy>();
-
-            private readonly Box background;
-
-            public CommentsHeader()
-            {
-                RelativeSizeAxes = Axes.X;
-                Height = height;
-                AddRangeInternal(new Drawable[]
-                {
-                    background = new Box
-                    {
-                        RelativeSizeAxes = Axes.Both,
-                    },
-                    new Container
-                    {
-                        RelativeSizeAxes = Axes.Both,
-                        Padding = new MarginPadding { Horizontal = padding },
-                        Children = new Drawable[]
-                        {
-                            new FillFlowContainer
-                            {
-                                AutoSizeAxes = Axes.Both,
-                                Direction = FillDirection.Horizontal,
-                                Spacing = new Vector2(spacing, 0),
-                                Anchor = Anchor.CentreLeft,
-                                Origin = Anchor.CentreLeft,
-                                Children = new Drawable[]
-                                {
-                                    new SpriteText
-                                    {
-                                        Font = OsuFont.GetFont(size: 14),
-                                        Text = @"Sort by"
-                                    }
-                                }
-                            }
-                        }
-                    }
-                });
-            }
-
-            [BackgroundDependencyLoader]
-            private void load(OsuColour colours)
-            {
-                background.Colour = colours.Gray4;
-            }
         }
     }
 }
